registry/client: split errors into new package
Split error handling functions from `registry/client` into a new
`registry/client/errors` package.

This allows consumers who are only interested in the error handling code
to import this package without getting big unrelated imports in their
binary. At the moment `/registry/client` will import
`/registry/storage/cache` -> `/metrics` -> `github.com/docker/go-metrics`
-> `github.com/prometheus/client_golang/prometheus` and so on.

The prometheus dependencies are big (about 700k in the binary) so it
would help us if we could remove them from the import chain.

This commit keeps backwards compatibility by type aliasing from the old
locations.
